Fix the incorrect way in which these objects were patched
So, an intersting thing happened when I attempted to build this project
on Travis CI - some of the tests failed with some really strange errors!

The problem here was that these tests were passing locally because the
environment was exactly what they expected, which led to the root cause:

These tests were mocked incorrectly :(

Live and learn!