Make sure that setup.py refers to all the charlesbot sub-packages
The problem I ran into here was that `pip install charlesbot` did not
bring in anything under the top `charlesbot` namespace -- so it was
missing `charlesbot.util`, etc.

While I was here I added a bit more metadata to setup.py and to make
sure it brings in its dependencies and that sort of thing.