Create a Slack singleton object responsible for the underlying socket connection + message interpreation
This simplifies the plugin architecture a bit as we don't have to pass
slack_client everywhere.

This also simplifies robot.py because will not need to deal with trying
to figure out what _kind_ of message object it just received.

Should note that this isn't a 'true' singleton, but rather the spirit of
a singleton object, but implemented using the Borg pattern.

Ref: http://www.aleax.it/5ep.html