Make the Help plugin a "special" plugin
This effectively acknowledges that the Help plugin is a plugin that
other plugins will (indirectly) need to interact with, and therefore
does not have to adhere to the strict plugin architecture.

What this actually does is dynamically loads and intializes the "help"
string from all enabled plugins, thus saving a person from having to
update the Help plugin manually.